support cmake3.18+ (#9):
To support NinJa, cmake 3.18+ changed generated defined flags for asm compiling. Specifically, it replaces "-D" with "--defsym" in flags.make. Dynamorio did not update for handling this change.

Without changing the compilation architecture of dynamorio, we fix this error by using a temp solution that is just running a script to replace "--defsym " to "-D" between cmake and make.